backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
gymnasium==0.29.1
numpy==1.24.3
scipy==1.11.2
torch==2.0.1
ray[rllib]==2.7.1
qiskit==0.44.1
qiskit-aer==0.12.2
pyspark==3.4.1
cassandra-driver==3.28.0
kafka-python==2.0.2
websockets==11.0.3
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
prometheus-client==0.17.1
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-fastapi==0.40b0
wandb==0.15.11
matplotlib==3.7.2
networkx==3.1
pandas==2.1.0
scikit-learn==1.3.0
plotly==5.16.1
streamlit==1.26.0
jupyter==1.0.0
notebook==7.0.3
requests==2.31.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1